Increased sequence limits from int to int64_t. Double max value is the limiting factor.

+++ src/Metrics/SSMetrics/Accuracy.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 Accuracy::Accuracy(const string& name, const string& refSequence, const string& predSequence, PrecalculatedMetric* precalculated) : Metric(name, refSequence, predSequence, precalculated) {
     for (const auto& secondaryStructure : GetSecondaryStructureClasses()) {
-        int summation = 0;
+        int64_t summation = 0;
         if (HasOverlappingBlocks(secondaryStructure)) {
             for (const auto& block : GetOverlappingBlocks(secondaryStructure)) {
                 summation += OverlapLength(block);
