+++ src/Metrics/SSMetrics/include/Regions/OverlapBlock.hpp
@@ -7,12 +7,12 @@
 class OverlapBlock : public Region
 {
 private:
-    int _length{0};
+    int64_t _length{0};
 public:
     SSBlock refRegion;
     SSBlock predRegion;
 
     OverlapBlock(const SSBlock& refRegion, const SSBlock& predRegion);
 
-    int GetLength() const;
+    int64_t GetLength() const;
 };
