+++ src/Metrics/SSMetrics/include/Sov94.hpp
@@ -8,7 +8,7 @@ private:
     bool _zeroDelta{false};
 
     bool GetZeroDelta() const;
-    int Delta(const OverlapBlock& overlapBlock);
+    int64_t Delta(const OverlapBlock& overlapBlock);
 public:
     Sov94(const string& name, const string& refSequence, const string& predSequence, const bool& zeroDelta, const bool& skipDuplicateRefBlocks, PrecalculatedMetric* precalculated);
