+++ src/Metrics/SSMetrics/include/Sov99.hpp
@@ -11,7 +11,7 @@ private:
     bool _zeroDelta{false};
 
     bool GetZeroDelta() const;
-    int Delta(const OverlapBlock& overlapBlock);
+    int64_t Delta(const OverlapBlock& overlapBlock);
 public:
     Sov99(const string& name, const string& refSequence, const string& predSequence, const bool& zeroDelta, PrecalculatedNormMetric* precalculatedNorm, PrecalculatedMetric* precalculated);
